Move ui component to ui folder
